Fetching project and separation between initial view and edit view
