+++ src/pages/project/new_project/EditProject.tsx
@@ -1,15 +1,17 @@
 import { Box } from '@mui/material';
 import { useTranslation } from 'react-i18next';
+import { useParams } from 'react-router-dom';
 import { PageHeader } from '../../../components/PageHeader';
 import { StyledPageContent } from '../../../components/styled/Wrappers';
-import StartEmptyProject from './StartEmptyProject';
+import { IdentifierParams } from '../../../utils/urlPaths';
+import { CreateProjectPage } from './CreateProjectPage';
+import ProjectForm from './ProjectForm';
 
 const EditProject = () => {
   const { t } = useTranslation();
+  const { identifier } = useParams<IdentifierParams>();
 
-  console.log('edit project');
-
-  return (
+  return !identifier ? (
     <StyledPageContent>
       <PageHeader>{t('project.create_project')}</PageHeader>
       <Box
@@ -19,8 +21,12 @@ const EditProject = () => {
           flexDirection: 'column',
           gap: '2rem',
         }}>
-        <StartEmptyProject onClose={() => {}} />
+        <CreateProjectPage onClose={() => {}} setInitialValues={() => {}} />
       </Box>
+    </StyledPageContent>
+  ) : (
+    <StyledPageContent>
+      <ProjectForm onClose={() => {}} identifier={decodeURIComponent(identifier)} />
     </StyledPageContent>
   );
 };
